MLOPS_Monitoring_v2/App/requirements.txt
fastapi
uvicorn[standard]
prometheus-client
openai==2.9.0
python-dotenv
httpx
numpy
pandas
fairlearn
scikit-learn